[nobug] Make sure tests clean up after themselves.
Many of the tests were reusing the same project metadata and
causing problems with other tests.  This commit makes sure that
the vast majority always start with a clean workspace and leave
a clean workspace when they are finished.  This helps issolate the
tests from one another.

Also files were being added to a static hashset that but were never
cleaned up after the Class under test was finished.  We reset this list
of items after the tests are done as well to a clean state and delete
any temporary files that were created.

Signed-off-by: David Carver <d_a_carver@yahoo.com>